refactor: Separate IEventRepository from IRepositories and enhance EventProcessorService logic
